Cover two missing states and improve the interface
There were two missing state cases I had failed to handle in the text >
unicode character substitutions. Both are exceptional cases so I went
with some warning-type characters.

I also improve the UI so that you don't have to edit the script. You
pass the NightScout URI to query as an command line argument instead.
This should make git-pull refreshes on the repo easier for anyone using
the script since they won't have to edit it to make it work.